requirement.txt
asgiref==3.4.1
boto3==1.18.22
botocore==1.21.22
Django==3.2.6
django-storages==1.11.1
djangorestframework==3.12.4
jmespath==0.10.0
python-dateutil==2.8.2
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
